redis_client.py
import redis
import os
from dotenv import load_dotenv

load_dotenv()

redis_client = redis.Redis(
    host=os.getenv("REDIS_HOST", "localhost"),
    port=int(os.getenv("REDIS_PORT", 6379)),
    decode_responses=True,
    socket_timeout=5,
    socket_connect_timeout=5
)

GEO_KEY = "drivers:locations"
